refactor: separate styles into different files
